Crate risc0_zkvm

The RISC Zero zkVM is a RISC-V virtual machine that produces zero-knowledge proofs of code it executes. By using the zkVM, a cryptographic receipt is produced which anyone can verify was produced by the zkVM’s guest code. No additional information about the code execution (such as, for example, the inputs provided) is revealed by publishing the receipt.

Additional (non-reference) resources for using our zkVM that you may also find helpful, especially if you’re new to the RISC Zero zkVM. These include:

Our zkVM Tutorial, which walks you through writing your first zkVM project.

The cargo risczero tool. It includes a new command which generates code for building and launching a zkVM guest and guidance on where projects most commonly modify host and guest code.

The examples, which contains various examples using our zkVM.

This clip from our presentation at ZK Hack III gives an overview of the RISC Zero zkVM. Our YouTube channel has many more videos as well.

We track zkVM issues with known workarounds using the rust guest workarounds GitHub tag. If you’re having problems running your code in the zkVM, you can see if there’s a workaround, and if you’re using a workaround, you can track when it gets resolved to a permanent solution.

And more on the RISC Zero developer website!

Crate Feature Flags

The following feature flags are supported.

Note that in order to use risc0-zkvm in the guest, you must disable the default features by setting default-features = false.

Feature	Target(s)	Implies	Description

client	all except rv32im	std	Enables the client API.

cuda		prove, std	Enables CUDA GPU acceleration for the prover. Requires CUDA toolkit to be installed.

disable-dev-mode	all except rv32im		Disables dev mode so that proving and verifying may not be faked. Used to prevent a misplaced RISC0_DEV_MODE from breaking security in production systems.

metal	macos	prove, std	Deprecated - Metal GPU acceleration for the prover is now enabled by default on Apple Silicon.

prove	all except rv32im	std	Enables the prover, incompatible within the zkvm guest.

std	all		Support for the Rust stdlib.

Modules

guest

The RISC Zero zkVM’s guest-side RISC-V API.

serde

Serialization and deserialization tools for the RISC Zero zkVM

sha

SHA-256 hashing services

Macros

declare_syscall

Construct a SyscallName declaration at compile time.

entry

Used for defining the guest’s entrypoint and main function.

Structs

ApiClient

A client implementation for interacting with a zkVM server.

Assumption

An assumption made in the course of proving program execution.

Assumptions

A list of assumptions, each a Digest or populated value of an Assumption.

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

CompositeReceipt

A receipt composed of one or more SegmentReceipt structs proving a single execution with continuations, and zero or more InnerAssumptionReceipt structs proving any assumptions.

CompositeReceiptVerifierParameters

Verifier parameters for CompositeReceipt.

ExecutorEnv

The Executor is configured from this object.

ExecutorEnvBuilder

A builder pattern used to construct an ExecutorEnv.

ExternalProver

An implementation of a Prover that runs proof workloads via an external r0vm process.

FakeReceipt

A fake receipt for testing and development.

Groth16Receipt

A receipt composed of a Groth16 over the BN_254 curve

Groth16ReceiptVerifierParameters

Verifier parameters used to verify a Groth16Receipt.

Groth16Seal

Groth16 seal object encoded in big endian.

Input

Input field in the ReceiptClaim, committing to a public value accessible to the guest.

InvalidExitCodeError

Error returned when a (system, user) exit code pair is an invalid representation.

Journal

A record of the public commitments for a proven zkVM execution.

Output

Output field in the ReceiptClaim, committing to a claimed journal and assumptions list.

ProveInfo

Information returned by the prover including receipt as well as other information useful for debugging

ProverOpts

Options to configure a Prover.

PrunedValueError

Error returned when the source value was pruned, and is not available.

Receipt

A receipt attesting to the execution of a guest program.

ReceiptClaim

Public claims about a zkVM guest execution, such as the journal committed to by the guest.

ReceiptMetadata

Metadata providing context on the receipt.

RedisParams

Determines the parameters for AssetRequest::Redis

SegmentInfo

Provides information about a segment of execution.

SegmentReceipt

A receipt attesting to the execution of a Segment.

SegmentReceiptVerifierParameters

Verifier parameters used to verify a SegmentReceipt.

SessionInfo

Provides information about the result of execution.

SessionStats

Struct containing information about a prover’s cycle count after running the guest program

SuccinctReceipt

A succinct receipt, produced via recursion, proving the execution of the zkVM with a STARK.

SuccinctReceiptVerifierParameters

Verifier parameters used to verify a SuccinctReceipt.

SystemState

Represents the public state of a segment, needed for continuations and receipt verification.

VerifierContext

Context available to the verification process.

Enums

Asset

Determines the format of an asset.

AssetRequest

Determines the format of an asset request.

AssumptionReceipt

An assumption attached to a guest execution as a result of calling env::verify or env::verify_integrity.

ExitCode

Exit condition indicated by the zkVM at the end of the guest execution.

InnerAssumptionReceipt

An enumeration of receipt types similar to InnerReceipt, but for use in AssumptionReceipt. Instead of proving only RISC-V execution with ReceiptClaim, this type can prove any claim implemented by one of its inner types.

InnerReceipt

A lower level receipt, containing the cryptographic seal (i.e. zero-knowledge proof) and verification logic for a specific proof system and circuit. All inner receipt types are zero-knowledge proofs of execution for a RISC-V zkVM.

MaybePruned

Either a source value or a hash Digest of the source value.

ReceiptKind

An enumeration of receipt kinds that can be requested to be generated.

TraceEvent

An event traced from the running VM.

Unknown

A type representing an unknown claim type.

Constants

ALLOWED_CONTROL_IDS

Control IDs allowed in the default set of recursion programs. Includes control IDs for the base set of recursion programs, and each power-of-two of the rv32im circuit, using Poseidon2.

ALLOWED_CONTROL_ROOT

Root of the Merkle tree constructed from ALLOWED_CONTROL_IDS, using Poseidon2.

DEFAULT_MAX_PO2

Maximum segment size, as a power of two (po2) that the default verifier parameters will accept.

GUEST_MAX_MEM

PAGE_SIZE

Size of a zkVM memory page.

POSEIDON2_CONTROL_IDS

Control IDs for each power-of-two of the rv32im circuit using Poseidon2.

VERSION

Reports the current version of this crate.

Traits

Connector

Connects a zkVM client and server

Executor

An Executor can execute a given ELF binary.

Prover

A Prover can execute a given ELF binary and produce a Receipt that can be used to verify correct computation.

TraceCallback

A callback used to collect TraceEvents.

Functions

align_up

Align address upwards.

compute_image_id

Compute and return the ImageID of the specified ELF binary.

default_executor

Return a default Executor based on environment variables and feature flags.

default_prover

Return a default Prover based on environment variables and feature flags.

get_version

Reports the current version of this crate as represented by a semver::Version.

is_dev_mode

Returns true if dev mode is enabled.

Type Aliases

Result

Result<T, Error>