risc0_zkvm::guest
Module env

Functions for interacting with the host environment.

The zkVM provides a set of functions to perform operations that manage execution, I/O, and proof composition. The set of functions related to each of these operations are described below.

System State

The guest has some control over the execution of the zkVM by pausing or exiting the program explicitly. This can be achieved using the pause and exit functions.

Proof Verification

The zkVM supports verification of RISC Zero receipts in a guest program, enabling proof composition. This can be achieved using the verify() and verify_integrity functions.

Input and Output

The zkVM provides a set of functions for handling input, public output, and private output. This is useful when interacting with the host and committing to some data publicly.

The zkVM provides functions that automatically perform (de)serialization on types and, for performance reasons, there is also a _slice variant that works with raw slices of plain old data. Performing operations on slices is more efficient, saving cycles during execution and consequently producing smaller proofs that are faster to produce. However, the _slice variants can be less ergonomic, so consider trade-offs when choosing between the two. For more information about guest optimization, see RISC Zero’s instruction on guest optimization

Convenience functions to read and write to default file descriptors are provided. See read(), write(), self::commit (and their _slice variants) for more information.

In order to access default file descriptors directly, see stdin, stdout, stderr and journal. These file descriptors are either FdReader or FdWriter instances, which can be used to read from or write to the host. To read from or write into them, use the Read and Write traits.

WARNING: Specifying a file descriptor with the same value of a default file descriptor is not recommended and may lead to unexpected behavior. A list of default file descriptors can be found in the fileno module.

Utility

The zkVM provides utility functions to log messages to the debug console and to measure the number of processor cycles that have occurred since the guest began. These can be achieved using the log and cycle_count functions.

Structs

FdReader

Provides a FdReader which can read from any file descriptor

FdWriter

Provides a FdWriter which can write to any file descriptor.

Enums

VerifyIntegrityError

Error encountered during a call to verify_integrity.

Traits

Read

Reads and deserializes objects

Write

Serializes and writes objects.

Functions

commit

Serialize the given data and commit it to the journal.

commit_slice

Commit the given slice to the journal.

cycle_count

Return the number of processor cycles that have occurred since the guest began.

exit

Terminate execution of the zkVM.

input_digest

Read the input digest from the input commitment.

journal

Return a writer for the JOURNAL.

log

Print a message to the debug console.

pause

Pause the execution of the zkVM.

read

Read private data from the STDIN of the zkVM and deserializes it.

read_slice

Read a slice from the STDIN of the zkVM.

send_recv_slice

Exchanges slices of plain old data with the host.

stderr

Return a writer for STDERR.

stdin

Return a reader for the standard input

stdout

Return a writer for STDOUT.

syscall

Exchange data with the host.

verify

Verify there exists a receipt for an execution with image_id and journal.

verify_assumption

Verify that there exists a valid receipt with the specified claim digest and control root.

verify_integrity

Verify that there exists a valid receipt with the specified ReceiptClaim.

write

Serialize the given data and write it to the STDOUT of the zkVM.

write_slice

Write the given slice to the STDOUT of the zkVM.