risc0_zkvm::guest::env
Function send_recv_slice

pub fn send_recv_slice<T: Pod, U: Pod>(
    syscall_name: SyscallName,
    to_host: &[T],
) -> &'static [U]

Exchanges slices of plain old data with the host.

This makes two calls to the given syscall; the first gets the length of the buffer to allocate for the return data, and the second actually receives the return data.

On the host side, implement SliceIo to provide a handler for this call.

NOTE: This method never frees up the buffer memory storing the host’s response.