risc0_zkvm::guest
Macro entry

macro_rules! entry {
    ($path:path) => { ... };
}

Used for defining the guest’s entrypoint and main function.

When #![no_main] is used, the programs entrypoint and main function is left undefined. The entry macro is required to indicate the main function and link it to an entrypoint provided by the RISC Zero SDK.

When std is enabled, the entrypoint will be linked automatically and this macro is not required.

Example

#![no_main]
#![no_std]

risc0_zkvm::entry!(main);

fn main() { }