risc0_zkvm::guest::sha
Struct Block

pub struct Block(/* private fields */);

Input block to the SHA-256 hashing algorithm. SHA-256 consumes blocks in 512-bit (64-byte) chunks in a Merkle–Damgård construction.

Implementations

impl Block

pub fn as_words(&self) -> &[u32]

Returns a reference to the Block as a slice of words.

pub fn as_bytes(&self) -> &[u8]

Returns a reference to the Block as a slice of bytes.

pub fn as_mut_words(&mut self) -> &mut [u32]

Returns a mutable slice of words.

pub fn as_half_blocks(&self) -> (&Digest, &Digest)

Returns the Block as references to two half-blocks, with the same size are a SHA-256 digest.

pub fn as_mut_bytes(&mut self) -> &mut [u8]

Returns a mutable slice of bytes.