risc0_zkvm::serde
Trait WordRead

pub trait WordRead {
    // Required methods
    fn read_words(&mut self, words: &mut [u32]) -> Result<()>;
    fn read_padded_bytes(&mut self, bytes: &mut [u8]) -> Result<()>;
}

A reader for reading streams with serialized word-based data

Required Methods

fn read_words(&mut self, words: &mut [u32]) -> Result<()>

Fill the given buffer with words from input. Returns an error if EOF was encountered.

fn read_padded_bytes(&mut self, bytes: &mut [u8]) -> Result<()>

Fill the given buffer with bytes from input, and discard the padding up to the next word boundary. Returns an error if EOF was encountered.