risc0_zkvm
Module sha

SHA-256 hashing services

Support for SHA-256 hashing compatible with the on-circuit SHA acceleration available in the guest. This module can be used from both the host and the guest – it will detect the build environment and select the appropriate implementation.

Usage

We provide two interfaces for SHA hashing. The first interface is based on Rust Crypto wrappers and can be found in the rust_crypto module, along with documentation and usage notes.

The other interface is to directly use an implementation of the Sha256 trait defined in this module:

use risc0_zkvm::sha::{Impl, Sha256};

// Hash a u8 array
let data = [1_u8, 2, 5, 14];
let hash = Impl::hash_bytes(&data);

// Hash a String
let abc = String::from("abc");
let abc_hash = Impl::hash_bytes(&abc.as_bytes());

// Hash a Digest
let hash_hash = Impl::hash_bytes(&hash.as_bytes());
// Hashing can also be done with 32-bit words (matching the zkVM word size)
let hash_hash_words = Impl::hash_words(&hash.as_words());
// Hashing with bytes or words should not change the result
assert_eq!(hash_hash, hash_hash_words);

Modules

rust_crypto

Rust Crypto wrappers for the RISC0 Sha256 trait.

Structs

Block

Input block to the SHA-256 hashing algorithm. SHA-256 consumes blocks in 512-bit (64-byte) chunks in a Merkle–Damgård construction.

Digest

Digest represents the results of a hashing function. It is always 256 bits of storage although depending on the hash it may have additional structure (for example Poseidon’s output is actually composed of field elements). The storage is in u32’s in part to simplify alignment requirements, especially in the zkVM.

Impl

A CPU-based Sha256 implementation.

Constants

BLOCK_BYTES

Size of the Block representation in bytes.

BLOCK_WORDS

The number of words in the representation of a Block.

DIGEST_BYTES

Size of the Digest representation in bytes.

DIGEST_SHORTS

The number of shorts in the representation of a Digest.

DIGEST_WORDS

The number of words in the representation of a Digest.

WORD_SIZE

Size of a zkVM machine word in bytes. 4 bytes (i.e. 32 bits) as the zkVM is an implementation of the rv32im ISA.

Statics

SHA256_INIT

Standard SHA-256 initialization vector.

Traits

Digestible

Defines a collision resistant hash for the typed and structured data.

Sha256

An implementation of the SHA-256 hashing algorithm of FIPS 180-4.