risc0_zkvm::sha
Struct Digest

pub struct Digest(/* private fields */);

Digest represents the results of a hashing function. It is always 256 bits of storage although depending on the hash it may have additional structure (for example Poseidon’s output is actually composed of field elements). The storage is in u32’s in part to simplify alignment requirements, especially in the zkVM.

Implementations

impl Digest

pub const ZERO: Digest = _

Digest of all zeroes.

pub const fn new(data: [u32; 8]) -> Digest

Constant constructor

pub const fn from_bytes(bytes: [u8; 32]) -> Digest

Construct a digest from a array of bytes in a const context. Outside of const context, Digest::from is recommended.

pub fn as_words(&self) -> &[u32]

Returns a reference to the Digest as a slice of words.

pub fn as_bytes(&self) -> &[u8]

Returns a reference to the Digest as a slice of bytes.

pub fn as_mut_words(&mut self) -> &mut [u32]

Returns a mutable slice of words.

pub fn as_mut_bytes(&mut self) -> &mut [u8]

Returns a mutable slice of bytes.